--- App.js
@@ -1,10 +1,0 @@
-import './App.css';
-import React from "react";
-import Aboutpage  from './about';
-import Loginpage from './login';
-
-function App(props) {
-  return (<Loginpage/>,<Aboutpage/>)
-};
-
-export default App;
